Add input parser

+++ include/depth_map.h
@@ -5,12 +5,12 @@
 
 class Depth_map: public Shader {
 	private:
-		float max_depth;
+		double max_depth;
 		rgb background;
 		rgb foreground;
 		
 	public:
-		Depth_map(float m_=0, rgb b_=rgb(), rgb f_=rgb()) 
+		Depth_map(double m_=0, rgb b_=rgb(), rgb f_=rgb()) 
 			: max_depth(m_), background(b_), foreground(f_) {}
 
 		rgb color(const Ray& ray, const Scene& scene) const;
