+++ include/hitable.h
@@ -7,7 +7,7 @@
 class Material;
 
 struct hit_record {
-	float t;
+	double t;
 	point3 p;
 	vec3 normal;
 	std::shared_ptr<Material> material;
@@ -16,7 +16,7 @@ struct hit_record {
 class Hitable {
 	public:
 		virtual ~Hitable() {}
-		virtual bool hit(const Ray &ray, float t_min, float t_max, hit_record &rec) const = 0;
+		virtual bool hit(const Ray &ray, double t_min, double t_max, hit_record &rec) const = 0;
 
 };
