+++ include/metal.h
@@ -6,9 +6,9 @@
 class Metal : public Material {
 	private:
 		vec3 albedo;
-		float fuzziness;
+		double fuzziness;
 	public:
-		Metal(vec3 a_, float f_) : albedo(a_), fuzziness(f_) {}
+		Metal(vec3 a_, double f_) : albedo(a_), fuzziness(f_) {}
 		
 		bool scatter(const Ray& ray, const hit_record& rec, vec3& attenuation, Ray& scattered) const;
 };
