+++ include/recursive.h
@@ -6,10 +6,10 @@
 
 class Recursive: public Shader {
 	private:
-		float max_steps;
+		int max_steps;
 
 	public:
-		Recursive(float s_=0) : max_steps(s_) {} 
+		Recursive(int s_=0) : max_steps(s_) {} 
 			
 		rgb color(const Ray& ray, const Scene& scene) const;
 		rgb recursive_color(const Ray& ray, const Scene& scene, int step) const;
