+++ include/vec3.h
@@ -13,7 +13,7 @@ class vec3
 {
     public:
         // Aliases
-        typedef float value_type;
+        typedef double value_type;
         enum field_t : int { X=0, Y=1, Z=2, R=0, G=1, B=2 };
 
         // Members
