+++ src/blinn_phong.cpp
@@ -3,7 +3,7 @@
 rgb Blinn_phong::color(const Ray& ray, const Scene& scene) const {
 	hit_record rec, aux;
 
-    if (scene.hit(ray, 0, std::numeric_limits<float>::max(), rec)) {
+    if (scene.hit(ray, 0, std::numeric_limits<double>::max(), rec)) {
     	std::vector<std::shared_ptr<Light>> lights = scene.get_lights();
     	std::shared_ptr<BP_material> material = std::dynamic_pointer_cast<BP_material>(rec.material);
     	vec3 unit_normal = unit_vector(rec.normal);
@@ -13,10 +13,10 @@ rgb Blinn_phong::color(const Ray& ray, const Scene& scene) const {
     	for(int i = 0; i < lights.size(); i++) {
     		vec3 unit_light = unit_vector(-lights[i]->direction);
     		Ray r(rec.p + 0.0001*unit_normal, unit_light);
-            if(!shadow || !scene.hit(r, 0.0001, std::numeric_limits<float>::max(), aux)) {
+            if(!shadow || !scene.hit(r, 0.0001, std::numeric_limits<double>::max(), aux)) {
         		
-                float d = fmax(dot(unit_light, unit_normal), 0.0);
-    	 		float s = pow(fmax(dot(unit_normal, unit_vector(unit_light + unit_ray)), 0.0), 
+                double d = fmax(dot(unit_light, unit_normal), 0.0);
+    	 		double s = pow(fmax(dot(unit_normal, unit_vector(unit_light + unit_ray)), 0.0), 
     	   			material->get_alpha());
     	    	diffuse += material->get_kd() * d * lights[i]->diffuse_intensity;
     	    	specular += material->get_ks() * s * lights[i]->specular_intensity;
@@ -33,8 +33,8 @@ rgb Blinn_phong::color(const Ray& ray, const Scene& scene) const {
     }
     else {
         vec3 unit_ray = unit_vector(ray.get_direction());
-        float x_ratio = (unit_ray.x()+1.0)/2.0;
-        float y_ratio = (unit_ray.y()+1.0)/2.0;
+        double x_ratio = (unit_ray.x()+1.0)/2.0;
+        double y_ratio = (unit_ray.y()+1.0)/2.0;
 
         //colors the background with bilinear interpolation
         rgb left = lerp(y_ratio, scene.get_background_upper_left(), scene.get_background_lower_left());
