+++ src/depth_map.cpp
@@ -3,8 +3,8 @@
 rgb Depth_map::color(const Ray& ray, const Scene& scene) const {
 	hit_record rec;
 
-    if (scene.hit(ray, 0, std::numeric_limits<float>::max(), rec)) {
-        float length = (rec.p - ray.get_origin()).length();
+    if (scene.hit(ray, 0, std::numeric_limits<double>::max(), rec)) {
+        double length = (rec.p - ray.get_origin()).length();
         if(length > max_depth) {
             return background;
         }
