+++ src/recursive.cpp
@@ -10,7 +10,7 @@ rgb Recursive::recursive_color(const Ray& ray, const Scene& scene, int step) con
     if(step == 0) {
         return rgb(0, 0, 0);
     }
-    if (scene.hit(ray, 0.001, std::numeric_limits<float>::max(), rec)) {
+    if (scene.hit(ray, 0.001, std::numeric_limits<double>::max(), rec)) {
         vec3 attenuation;
         Ray scattered;
         if(rec.material->scatter(ray, rec, attenuation, scattered)) {
@@ -22,8 +22,8 @@ rgb Recursive::recursive_color(const Ray& ray, const Scene& scene, int step) con
     }
     else {
         vec3 unit_ray = unit_vector(ray.get_direction());
-        float x_ratio = (unit_ray.x()+1.0)/2.0;
-        float y_ratio = (unit_ray.y()+1.0)/2.0; 
+        double x_ratio = (unit_ray.x()+1.0)/2.0;
+        double y_ratio = (unit_ray.y()+1.0)/2.0; 
 
         //colors the background with bilinear interpolation
         rgb left = lerp(y_ratio, scene.get_background_upper_left(), scene.get_background_lower_left());
