+++ src/sphere.cpp
@@ -1,15 +1,15 @@
 #include "../include/sphere.h"
 
-bool Sphere::hit(const Ray &ray, float t_min, float t_max, hit_record &rec) const {
+bool Sphere::hit(const Ray &ray, double t_min, double t_max, hit_record &rec) const {
 	vec3 direction = ray.get_direction();
     vec3 center_direction = ray.get_origin() - center;
 
-    float a = dot(direction, direction);
-    float b = 2 * dot(center_direction, direction);
-    float c = dot(center_direction, center_direction) - (radius * radius);
-    float delta = b * b - 4 * a * c;
+    double a = dot(direction, direction);
+    double b = 2 * dot(center_direction, direction);
+    double c = dot(center_direction, center_direction) - (radius * radius);
+    double delta = b * b - 4 * a * c;
     if (delta >= 0) {
-        float t = (-b - sqrt(delta)) / (2 * a);
+        double t = (-b - sqrt(delta)) / (2.0 * a);
         if (t >= t_min && t <= t_max) {
         	rec.t = t;
         	rec.p = ray.point_at(t);
@@ -17,7 +17,7 @@ bool Sphere::hit(const Ray &ray, float t_min, float t_max, hit_record &rec) cons
             rec.material = material;
         	return true;
         }
-        t = (-b + sqrt(delta)) / (2 * a);
+        t = (-b + sqrt(delta)) / (2.0 * a);
         if (t >= t_min && t <= t_max) {
         	rec.t = t;
         	rec.p = ray.point_at(t);
